+++ apps/endringsmelding-pleiepenger/src/app/api/endpoints/søknadStateEndpoint.ts
@@ -1,6 +1,7 @@
+import { Søker } from '@navikt/sif-common-api';
 import persistence, { PersistenceInterface } from '@navikt/sif-common-core-ds/src/utils/persistence/persistence';
 import { jsonSort } from '@navikt/sif-common-utils';
-import { K9Sak, Søker, Søknadsdata, ValgteEndringer } from '@types';
+import { K9Sak, Søknadsdata, ValgteEndringer } from '@types';
 import { AxiosResponse } from 'axios';
 import hash from 'object-hash';
 import { APP_VERSJON } from '../../constants/APP_VERSJON';