+++ apps/pleiepenger-i-livets-sluttfase-soknad/src/app/api/useSøknadInitialData.ts
@@ -2,11 +2,10 @@ import { isUnauthorized } from '@navikt/sif-common-core-ds/src/utils/apiUtils';
 import { useEffect, useState } from 'react';
 import { SØKNAD_VERSJON } from '../constants/SØKNAD_VERSJON';
 import { RequestStatus } from '../types/RequestStatus';
-import { Søker } from '../types/Søker';
 import { SøknadContextState } from '../types/SøknadContextState';
 import { SøknadRoutes } from '../types/SøknadRoutes';
 import appSentryLogger from '../utils/appSentryLogger';
-import { fetchSøker } from '@navikt/sif-common-api';
+import { fetchSøker, Søker } from '@navikt/sif-common-api';
 import { MellomlagringData, mellomlagringService } from './mellomlagringService';
 
 export type SøknadInitialData = SøknadContextState;